Fix quotetarget
