+++ frame/runtime/src/traits.rs
@@ -11,7 +11,7 @@ use frame_common::{
     traits::*,
 };
 use frame_treekem::{handshake::HandshakeParams, DhPubKey, EciesCiphertext};
-use remote_attestation::RAService;
+use remote_attestation::Quote;
 
 /// Execute state transition functions from runtime
 pub trait RuntimeExecutor<G: ContextOps>: Sized {
@@ -113,5 +113,5 @@ pub trait QuoteGetter: Sized {
     /// Generate Base64-encoded QUOTE data structure.
     /// QUOTE will be sent to Attestation Service to verify SGX's status.
     /// For more information: https://api.trustedservices.intel.com/documents/sgx-attestation-api-spec.pdf
-    fn quote(&self) -> Result<RAService>;
+    fn quote(&self) -> Result<Quote>;
 }
